Re-arrange functions
