+++ pycontrails/models/cocip/output/grid_cirrus.py
@@ -16,378 +16,7 @@ from pycontrails.physics.geo import spatial_bounding_box
 np.random.seed(1)
 
 
-def contrails_to_hi_res_grid(
-        time: pd.Timestamp | np.datetime64,
-        contrails_t: GeoVectorDataset, *,
-        var_name: str,
-        spatial_bbox: list[float] = [-180, -90, 180, 90],
-        spatial_grid_res: float = 0.05,
-) -> xr.DataArray:
-    """
-    Aggregate contrail segments to a high-resolution longitude-latitude grid.
 
-    Parameters
-    ----------
-    time : pd.Timestamp | np.datetime64
-        UTC time of interest.
-    contrails_t : GeoVectorDataset
-        All contrail waypoint outputs at `time`.
-    var_name : str
-        Contrail property for aggregation, where `var_name` must be included in `contrail_segment`.
-        For example, `tau_contrail`, `rf_sw`, `rf_lw`, and `rf_net`
-    spatial_bbox : list[float]
-        Spatial bounding box, [lon_min, lat_min, lon_max, lat_max], [:math:`\deg`]
-    spatial_grid_res : float
-        Spatial grid resolution, [:math:`\deg`]
-
-    Returns
-    -------
-    xr.DataArray
-        Contrail segments and their properties aggregated to a longitude-latitude grid.
-    """
-    # Ensure the required columns are included in `contrails_t`
-    cols_req = [
-        "flight_id", "waypoint", "longitude", "latitude",
-        "altitude", "time", "sin_a", "cos_a", "width", var_name
-    ]
-    contrails_t.ensure_vars(cols_req)
-
-    # Ensure that the times in `contrails_t` are the same.
-    is_in_time = contrails_t["time"] == time
-    if ~np.all(is_in_time):
-        warnings.warn(
-            f"Contrails have inconsistent times. Waypoints that are not in {time} are removed."
-        )
-        contrails_t = contrails_t.filter(is_in_time)
-
-    main_grid = _initialise_longitude_latitude_grid(spatial_bbox, spatial_grid_res)
-
-    # Contrail head and tails: continuous segments only
-    heads_t = contrails_t.dataframe
-    heads_t.sort_values(["flight_id", "waypoint"], inplace=True)
-    tails_t = heads_t.shift(periods=-1)
-
-    is_continuous = heads_t["continuous"]
-    heads_t = heads_t[is_continuous].copy()
-    tails_t = tails_t[is_continuous].copy()
-    tails_t["waypoint"] = tails_t['waypoint'].astype('int')
-
-    heads_t.set_index(["flight_id", "waypoint"], inplace=True, drop=False)
-    tails_t.index = heads_t.index
-
-    # Aggregate contrail segments to a high resolution longitude-latitude grid
-    for i in tqdm(heads_t.index[:2000]):
-        contrail_segment = GeoVectorDataset(
-            pd.concat([heads_t[cols_req].loc[i], tails_t[cols_req].loc[i]], axis=1).T,
-            copy=True
-        )
-
-        segment_grid = segment_property_to_hi_res_grid(
-            contrail_segment, var_name=var_name, spatial_grid_res=spatial_grid_res
-        )
-        main_grid = _add_segment_to_main_grid(main_grid, segment_grid)
-
-    return main_grid
-
-
-def _initialise_longitude_latitude_grid(
-        spatial_bbox: list[float] = [-180, -90, 180, 90],
-        spatial_grid_res: float = 0.05,
-) -> xr.DataArray:
-    """
-    Create longitude-latitude grid of specified coordinates and spatial resolution.
-
-    Parameters
-    ----------
-    spatial_bbox : list[float]
-        Spatial bounding box, [lon_min, lat_min, lon_max, lat_max], [:math:`\deg`]
-    spatial_grid_res : float
-        Spatial grid resolution, [:math:`\deg`]
-
-    Returns
-    -------
-    xr.DataArray
-        Longitude-latitude grid of specified coordinates and spatial resolution, filled with zeros.
-
-    Notes
-    -----
-    This empty grid is used to store the aggregated contrail properties of the individual
-    contrail segments, such as the gridded contrail optical depth and radiative forcing.
-    """
-    lon_coords = np.arange(
-        spatial_bbox[0], spatial_bbox[2] + spatial_grid_res, spatial_grid_res
-    )
-    lat_coords = np.arange(
-        spatial_bbox[1], spatial_bbox[3] + spatial_grid_res, spatial_grid_res
-    )
-    return xr.DataArray(
-        np.zeros((len(lon_coords), len(lat_coords))),
-        dims=["longitude", "latitude"],
-        coords={"longitude": lon_coords, "latitude": lat_coords},
-    )
-
-
-def segment_property_to_hi_res_grid(
-        contrail_segment: GeoVectorDataset, *,
-        var_name: str,
-        spatial_grid_res: float = 0.05,
-) -> xr.DataArray:
-    """
-    Convert the contrail segment property to a high-resolution longitude-latitude grid.
-
-    Parameters
-    ----------
-    contrail_segment : GeoVectorDataset
-        Contrail segment waypoints (head and tail).
-    var_name : str
-        Contrail property of interest, where `var_name` must be included in `contrail_segment`.
-        For example, `tau_contrail`, `rf_sw`, `rf_lw`, and `rf_net`
-    spatial_grid_res : float
-        Spatial grid resolution, [:math:`\deg`]
-
-    Returns
-    -------
-    xr.DataArray
-        Contrail segment dimension and property projected to a longitude-latitude grid.
-
-    Notes
-    -----
-    - See Appendix A11 and A12 of :cite:`schumannContrailCirrusPrediction2012`.
-    """
-    # Ensure that `contrail_segment` contains the required variables
-    contrail_segment.ensure_vars(("sin_a", "cos_a", "width", var_name))
-
-    # Ensure that `contrail_segment` only contains two waypoints and have the same time.
-    assert len(contrail_segment) == 2
-    assert contrail_segment["time"][0] == contrail_segment["time"][1]
-
-    # Calculate contrail edges
-    (
-        contrail_segment["lon_edge_l"],
-        contrail_segment["lat_edge_l"],
-        contrail_segment["lon_edge_r"],
-        contrail_segment["lat_edge_r"],
-    ) = contrail_properties.contrail_edges(
-        contrail_segment["longitude"],
-        contrail_segment["latitude"],
-        contrail_segment["sin_a"],
-        contrail_segment["cos_a"],
-        contrail_segment["width"],
-    )
-
-    # Initialise contrail segment grid with spatial domain that covers the contrail area.
-    lon_edges = np.concatenate(
-        [contrail_segment["lon_edge_l"], contrail_segment["lon_edge_r"]], axis=0
-    )
-    lat_edges = np.concatenate(
-        [contrail_segment["lat_edge_l"], contrail_segment["lat_edge_r"]], axis=0
-    )
-    spatial_bbox = spatial_bounding_box(lon_edges, lat_edges, buffer=0.5)
-    segment_grid = _initialise_longitude_latitude_grid(spatial_bbox, spatial_grid_res)
-
-    # Calculate gridded contrail segment properties
-    weights = _pixel_weights(contrail_segment, segment_grid)
-    dist_perpendicular = _segment_perpendicular_distance_to_pixels(contrail_segment, weights)
-    plume_concentration = _gaussian_plume_concentration(
-        contrail_segment, weights, dist_perpendicular
-    )
-
-    # Distribute selected contrail property to grid
-    return plume_concentration * (
-            weights * xr.ones_like(weights) * contrail_segment[var_name][1]
-            + (1 - weights) * xr.ones_like(weights) * contrail_segment[var_name][0]
-    )
-
-
-def _pixel_weights(
-        contrail_segment: GeoVectorDataset,
-        segment_grid: xr.DataArray
-) -> xr.DataArray:
-    """
-    Calculate the pixel weights for `segment_grid`.
-
-    Parameters
-    ----------
-    contrail_segment : GeoVectorDataset
-        Contrail segment waypoints (head and tail).
-    segment_grid : xr.DataArray
-        Contrail segment grid with spatial domain that covers the contrail area.
-
-    Returns
-    -------
-    xr.DataArray
-        Pixel weights for `segment_grid`
-
-    Notes
-    -----
-    - See Appendix A12 of :cite:`schumannContrailCirrusPrediction2012`.
-    - This is the weights (from the beginning of the contrail segment) to the nearest longitude and
-        latitude pixel in the `segment_grid`.
-    - The contrail segment do not contribute to the pixel if weight < 0 or > 1.
-    """
-    head = contrail_segment.dataframe.iloc[0]
-    tail = contrail_segment.dataframe.iloc[1]
-
-    # Calculate determinant
-    dx = units.longitude_distance_to_m(
-        (tail["longitude"] - head["longitude"]),
-        0.5 * (head["latitude"] + tail["latitude"]),
-    )
-    dy = units.latitude_distance_to_m(tail["latitude"] - head["latitude"])
-    det = dx**2 + dy**2
-
-    # Calculate pixel weights
-    lon_grid, lat_grid = np.meshgrid(
-        segment_grid["longitude"].values, segment_grid["latitude"].values
-    )
-    dx_grid = units.longitude_distance_to_m(
-        (lon_grid - head["longitude"]),
-        0.5 * (head["latitude"] + lat_grid),
-    )
-    dy_grid = units.latitude_distance_to_m((lat_grid - head["latitude"]))
-    weights = (dx * dx_grid + dy * dy_grid) / det
-    return xr.DataArray(
-        data=weights.T,
-        dims=["longitude", "latitude"],
-        coords={"longitude": segment_grid["longitude"], "latitude": segment_grid["latitude"]},
-    )
-
-
-def _segment_perpendicular_distance_to_pixels(
-        contrail_segment: GeoVectorDataset,
-        weights: xr.DataArray
-) -> xr.DataArray:
-    """
-    Calculate perpendicular distance from contrail segment to each segment grid pixel.
-
-    Parameters
-    ----------
-    contrail_segment : GeoVectorDataset
-        Contrail segment waypoints (head and tail).
-    weights : xr.DataArray
-        Pixel weights for `segment_grid`.
-        See `_pixel_weights` function.
-
-    Returns
-    -------
-    xr.DataArray
-        Perpendicular distance from contrail segment to each segment grid pixel, [:math:`m`]
-
-    Notes
-    -----
-    - See Figure A7 of :cite:`schumannContrailCirrusPrediction2012`.
-    """
-    head = contrail_segment.dataframe.iloc[0]
-    tail = contrail_segment.dataframe.iloc[1]
-
-    # Longitude and latitude along contrail segment
-    lon_grid, lat_grid = np.meshgrid(
-        weights["longitude"].values, weights["latitude"].values
-    )
-
-    lon_s = head["longitude"] + weights.T.values * (tail["longitude"] - head["longitude"])
-    lat_s = head["latitude"] + weights.T.values * (tail["latitude"] - head["latitude"])
-
-    lon_dist = units.longitude_distance_to_m(
-        np.abs(lon_grid - lon_s),
-        0.5 * (lat_s + lat_grid)
-    )
-
-    lat_dist = units.latitude_distance_to_m(np.abs(lat_grid - lat_s))
-    dist_perp = (lon_dist**2 + lat_dist**2) ** 0.5
-    return xr.DataArray(dist_perp.T, coords=weights.coords)
-
-
-def _gaussian_plume_concentration(
-        contrail_segment: GeoVectorDataset,
-        weights: xr.DataArray,
-        dist_perpendicular: xr.DataArray,
-) -> xr.DataArray:
-    """
-    Calculate relative gaussian plume concentration along the contrail width.
-
-    Parameters
-    ----------
-    contrail_segment : GeoVectorDataset
-        Contrail segment waypoints (head and tail).
-    weights : xr.DataArray
-        Pixel weights for `segment_grid`.
-        See `_pixel_weights` function.
-    dist_perpendicular : xr.DataArray
-        Perpendicular distance from contrail segment to each segment grid pixel, [:math:`m`]
-        See `_segment_perpendicular_distance_to_pixels` function.
-
-    Returns
-    -------
-    xr.DataArray
-        Relative gaussian plume concentration along the contrail width
-
-    Notes
-    -----
-    - Assume a one-dimensional Gaussian plume.
-    - See Appendix A11 of :cite:`schumannContrailCirrusPrediction2012`.
-    """
-    head = contrail_segment.dataframe.iloc[0]
-    tail = contrail_segment.dataframe.iloc[1]
-
-    width = weights.values * tail["width"] + (1 - weights.values) * head["width"]
-    sigma_yy = 0.125 * width**2
-
-    concentration = np.where(
-        (weights.values < 0) | (weights.values > 1),
-        0,
-        (4 / np.pi)**0.5 * np.exp(-0.5 * dist_perpendicular.values**2 / sigma_yy)
-    )
-    return xr.DataArray(concentration, coords=weights.coords)
-
-
-def _add_segment_to_main_grid(
-        main_grid: xr.DataArray,
-        segment_grid: xr.DataArray
-) -> xr.DataArray:
-    """
-    Add the gridded contrail segment to the main grid.
-
-    Parameters
-    ----------
-    main_grid : xr.DataArray
-        Aggregated contrail segment properties in a longitude-latitude grid.
-    segment_grid : xr.DataArray
-        Contrail segment dimension and property projected to a longitude-latitude grid.
-
-    Returns
-    -------
-    xr.DataArray
-        Aggregated contrail segment properties, including `segment_grid`.
-
-    Notes
-    -----
-    - The spatial domain of `segment_grid` only covers the contrail segment, which is added to
-        the `main_grid` which is expected to have a larger spatial domain than the `segment_grid`.
-    - This architecture is used to reduce the computational resources.
-    """
-    lon_main = main_grid["longitude"].values
-    lat_main = main_grid["latitude"].values
-
-    lon_segment_grid = np.round(segment_grid["longitude"].values, decimals=2)
-    lat_segment_grid = np.round(segment_grid["latitude"].values, decimals=2)
-
-    main_grid_arr = main_grid.values
-    subgrid_arr = segment_grid.values
-
-    try:
-        ix_ = np.searchsorted(lon_main, lon_segment_grid[0])
-        ix = np.searchsorted(lon_main, lon_segment_grid[-1]) + 1
-        iy_ = np.searchsorted(lat_main, lat_segment_grid[0])
-        iy = np.searchsorted(lat_main, lat_segment_grid[-1]) + 1
-    except IndexError:
-        warnings.warn(
-            "Contrail segment ignored as it is outside spatial bounding box of the main grid. "
-        )
-    else:
-        main_grid_arr[ix_:ix, iy_:iy] = main_grid_arr[ix_:ix, iy_:iy] + subgrid_arr
-
-    return xr.DataArray(main_grid_arr, coords=main_grid.coords)
 
 
 # ---
